refactor: ♻️ (user) incrementPointをUserGamaManagerUseCaseに移動
