+++ src/module/user/domain/service/use-case/user-game-manager.use-case.ts
@@ -1,8 +1,10 @@
 import { User } from '../../model/user.model';
 import { ExitGameArgs } from '~/user/controller/dto/args/exit-game.args';
+import { IncrementPointArgs } from '~/user/controller/dto/args/increment-point.args';
 import { JoinGameArgs } from '~/user/controller/dto/args/join-game.args';
 
 export interface UserGameManagerUseCaseInterface {
   joinGame(args: JoinGameArgs): Promise<User>;
   exitGame(args: ExitGameArgs): Promise<User>;
+  incrementPoint(args: IncrementPointArgs, isBeforeDay2: boolean): Promise<User[]>;
 }
