+++ src/module/user/domain/service/use-case/user-updater.use-case.ts
@@ -1,8 +1,6 @@
 import { User } from '../../model/user.model';
-import { IncrementPointArgs } from '~/user/controller/dto/args/increment-point.args';
 import { UpdateUserArgs } from '~/user/controller/dto/args/update-user.args';
 
 export interface UserUpdaterUseCaseInterface {
   updateUser(args: UpdateUserArgs): Promise<User>;
-  incrementPoint(args: IncrementPointArgs, isBeforeDay2: boolean): Promise<User[]>;
 }
