+++ src/module/user/use-case/user-updater.use-case.ts
@@ -1,8 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { Game } from '@prisma/client';
-import { IncrementPointArgs } from '../controller/dto/args/increment-point.args';
 import { UpdateUserArgs } from '../controller/dto/args/update-user.args';
-import { User } from '../domain/model/user.model';
 import { UserRepositoryInterface } from '../domain/service/repository/user.repository';
 import { UserUpdaterUseCaseInterface } from '../domain/service/use-case/user-updater.use-case';
 import { InjectionToken } from '@/common/constant/injection-token.constant';
@@ -18,45 +15,5 @@ export class UserUpdaterUseCase implements UserUpdaterUseCaseInterface {
     const updatedUser = await this.userRepository.update(args);
 
     return updatedUser;
-  }
-
-  async incrementPoint(args: IncrementPointArgs, isBeforeDay2: boolean) {
-    const updatedUsers = await Promise.all(
-      args.users.map(async (arg) => {
-        const foundUser = await this.userRepository.findUnique({
-          where: { id: arg.id },
-        });
-        if (!foundUser) {
-          throw new Error('User not found');
-        }
-
-        let updatedUser: User;
-        if (isBeforeDay2) {
-          updatedUser = await this.userRepository.update({
-            where: { id: arg.id },
-            data: {
-              totalPointDay1: foundUser.totalPointDay1 + arg.increment,
-              consumablePoint: foundUser.consumablePoint + arg.increment,
-              pullableGachaTimes: foundUser.pullableGachaTimes + 1,
-              participateGame: Game.NONE,
-            },
-          });
-        } else {
-          updatedUser = await this.userRepository.update({
-            where: { id: arg.id },
-            data: {
-              totalPointDay2: foundUser.totalPointDay2 + arg.increment,
-              consumablePoint: foundUser.consumablePoint + arg.increment,
-              pullableGachaTimes: foundUser.pullableGachaTimes + 1,
-              participateGame: Game.NONE,
-            },
-          });
-        }
-
-        return updatedUser;
-      }),
-    );
-
-    return updatedUsers;
   }
 }
